Adapt imports
