+++ gammapy/datasets/core.py
@@ -5,8 +5,8 @@ import copy
 import numpy as np
 from gammapy.utils.scripts import make_path, read_yaml, write_yaml
 from gammapy.utils.table import table_from_row_data
-from ..maps import WcsNDMap
-from .parameter import Parameters
+from gammapy.maps import WcsNDMap
+from gammapy.modeling.parameter import Parameters
 
 __all__ = ["Dataset", "Datasets"]
 
@@ -158,7 +158,7 @@ class Datasets(collections.abc.Sequence):
         dataset : 'gammapy.modeling.Datasets'
             Datasets
         """
-        from .serialize import dict_to_datasets
+        from gammapy.modeling.serialize import dict_to_datasets
 
         components = read_yaml(make_path(filemodel))
         data_list = read_yaml(make_path(filedata))
@@ -177,7 +177,7 @@ class Datasets(collections.abc.Sequence):
         overwrite : bool
             overwrite datasets FITS files
         """
-        from .serialize import datasets_to_dict
+        from gammapy.modeling.serialize import datasets_to_dict
 
         path = make_path(path)
