+++ gammapy/modeling/__init__.py
@@ -1,6 +1,5 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 """Models and fitting."""
-from .datasets import *
 from .fit import *
 from .model import *
 from .parameter import *
