+++ gammapy/modeling/tests/test_fit.py
@@ -2,35 +2,13 @@
 """Unit tests for the Fit class"""
 import pytest
 from numpy.testing import assert_allclose
-from gammapy.modeling import Fit, Parameter, Parameters
+from gammapy.modeling import Fit
 from gammapy.utils.testing import requires_dependency
+from gammapy.datasets.tests.test_core import MyDataset
 
 pytest.importorskip("iminuit")
 
 
-class MyDataset:
-    def __init__(self, name=""):
-        self.name = name
-        self.parameters = Parameters(
-            [Parameter("x", 2), Parameter("y", 3e2), Parameter("z", 4e-2)]
-        )
-        self.data_shape = (1,)
-
-    def stat_sum(self):
-        # self._model.parameters = parameters
-        x, y, z = [p.value for p in self.parameters]
-        x_opt, y_opt, z_opt = 2, 3e2, 4e-2
-        return (x - x_opt) ** 2 + (y - y_opt) ** 2 + (z - z_opt) ** 2
-
-    def fcn(self):
-        x, y, z = [p.value for p in self.parameters]
-        x_opt, y_opt, z_opt = 2, 3e5, 4e-5
-        x_err, y_err, z_err = 0.2, 3e4, 4e-6
-        return (
-            ((x - x_opt) / x_err) ** 2
-            + ((y - y_opt) / y_err) ** 2
-            + ((z - z_opt) / z_err) ** 2
-        )
 
 
 @pytest.mark.parametrize("backend", ["minuit"])
