+++ gammapy/spectrum/tests/test_flux_point.py
@@ -5,7 +5,8 @@ from numpy.testing import assert_allclose
 import astropy.units as u
 from astropy.table import Table
 from gammapy.catalog.fermi import SourceCatalog3FGL
-from gammapy.modeling import Datasets, Fit
+from gammapy.modeling import Fit
+from gammapy.datasets import Datasets
 from gammapy.modeling.models import PowerLawSpectralModel, SkyModel, SpectralModel
 from gammapy.spectrum import FluxPoints, FluxPointsDataset
 from gammapy.utils.testing import (
